Save every single footprint using kicad v6
During the development the file format for footprints changed a little. Most
important, items are not sorted and have a timestamp.
To make future changes to those files easier to review (e.g. the diff only
shows the lines with changes), all files need to be converted to the latest
format.
Assuming there are no bugs in the footprint load/save functionality, this
commit does not have any 'real' changes. All the footprints should render
the same.