Add HD15 Dsub connector with 11.43mm hole edge spacing.
The existing part with a 10.89mm hole spacing does not seem to
match any standard dsub connector I have been able to find, I'm
not sure if it was a mistake originally or there really was a part
with that spacing. If it really was a mistake, it might make sense
to delete the footprint.

The datasheet I used is:

http://www.assmann-wsw.com/uploads/datasheets/AHDxxxA-KG-TAxB.pdf

I based this off the existing footprint, moved the holes, extended
the courtyard and silkscreen, and re-centered on pin 1.

I do not know how to update the 3D model so I removed it for now.